Fixed error in random role distribution.
Bugfix: The same role was assigned to everyone in the match due to
a typo in PHP variable.

Bugfix: match duration was 5 minutes regardless of the number
of players instead of being n+5 if n>5 (typo in PHP variable)

Bugfix: Whatsapp link was not working as it was overwritten
by localization.

Bugfix: if the 'first' player was stroken out on the list it
would come back "normal" at the next update of the page be-
cause the "(first)" string was interpreted as part of the name.

Added translation for the cookie banner.