Added kick mechanism.
The host can now kick any player during the non-active phase of the
game (while waitng for everyone to connect and the game is not yet
going on).
The host cannot kick himself out of the match as this would result
in a loss of control over the match itself.
Kicked players are redirected to the homepage.

The page is now updated every 5 seconds.

Bugfix: hidden player data was shown everytime the getUpdate()
function was called.

Bugfix: new players may now join the match correctly even if typing
the key with capital letters.